replace old linear downloader
