+++ crates/net/downloaders/src/headers/mod.rs
@@ -1,3 +1,2 @@
-pub mod concurrent;
 /// A Linear downloader implementation.
 pub mod linear;
