+++ crates/stages/src/stages/headers.rs
@@ -16,7 +16,7 @@ use reth_interfaces::{
         error::DownloadError,
         headers::{
             client::{HeadersClient, StatusUpdater},
-            downloader::{ensure_parent, HeaderDownloader2},
+            downloader::{ensure_parent, HeaderDownloader},
         },
     },
 };
@@ -40,7 +40,7 @@ pub(crate) const HEADERS: StageId = StageId("Headers");
 /// NOTE: This stage downloads headers in reverse. Upon returning the control flow to the pipeline,
 /// the stage progress is not updated unless this stage is done.
 #[derive(Debug)]
-pub struct HeaderStage<D: HeaderDownloader2, C: Consensus, H: HeadersClient, S: StatusUpdater> {
+pub struct HeaderStage<D: HeaderDownloader, C: Consensus, H: HeadersClient, S: StatusUpdater> {
     /// Strategy for downloading the headers
     pub downloader: D,
     /// Consensus client implementation
@@ -56,7 +56,7 @@ pub struct HeaderStage<D: HeaderDownloader2, C: Consensus, H: HeadersClient, S: 
 }
 
 #[async_trait::async_trait]
-impl<DB: Database, D: HeaderDownloader2, C: Consensus, H: HeadersClient, S: StatusUpdater> Stage<DB>
+impl<DB: Database, D: HeaderDownloader, C: Consensus, H: HeadersClient, S: StatusUpdater> Stage<DB>
     for HeaderStage<D, C, H, S>
 {
     /// Return the id of the stage
@@ -152,7 +152,7 @@ impl<DB: Database, D: HeaderDownloader2, C: Consensus, H: HeadersClient, S: Stat
     }
 }
 
-impl<D: HeaderDownloader2, C: Consensus, H: HeadersClient, S: StatusUpdater>
+impl<D: HeaderDownloader, C: Consensus, H: HeadersClient, S: StatusUpdater>
     HeaderStage<D, C, H, S>
 {
     async fn update_head<DB: Database>(
